test: default migrations order
