+++ app/module/manager.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"slices"
-	"sort"
 
 	sdkmodule "github.com/cosmos/cosmos-sdk/types/module"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -360,33 +359,4 @@ func (m *Manager) assertNoForgottenModules(setOrderFnName string, moduleNames []
 		panic(fmt.Sprintf(
 			"%s: all modules must be defined when setting %s, missing: %v", setOrderFnName, setOrderFnName, missing))
 	}
-}
-
-// defaultMigrationsOrder returns a default migrations order: ascending alphabetical by module name,
-// except x/auth which will run last, see:
-// https://github.com/cosmos/cosmos-sdk/issues/10591
-func defaultMigrationsOrder(modules []string) []string {
-	const authName = "auth"
-	out := make([]string, 0, len(modules))
-	hasAuth := false
-	for _, m := range modules {
-		if m == authName {
-			hasAuth = true
-		} else {
-			out = append(out, m)
-		}
-	}
-	sort.Strings(out)
-	if hasAuth {
-		out = append(out, authName)
-	}
-	return out
-}
-
-func getKeys(m map[uint64]map[string]sdkmodule.AppModule) []uint64 {
-	keys := make([]uint64, 0, len(m))
-	for key := range m {
-		keys = append(keys, key)
-	}
-	return keys
 }
